+++ src/main/java/no/nav/pam/geography/CountryService.java
@@ -1,7 +1,6 @@
 package no.nav.pam.geography;
 
 
-import no.nav.pam.geography.model.Country;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
